Don't include image digest in tag
For now, we are doing this because we want to avoid unnecessary Flux
reconciliations after deployments on local environments. We will
carefully re-examine if it is worth it reintroducing the digest or the
commit ID in the tag. This will help in easily knowing which version of
an image is deployed or to which commit ID an image is related. Also,
every deployment will be reflected in the Helm chart's commit history.